Remove all uses of GUI score director
